Move existing login test to correct file
