+++ tests/Feature/Http/Controllers/Auth/LoginControllerTest.php
@@ -2,7 +2,6 @@
 
 namespace Tests\Feature\Http\Controllers\Auth;
 
-use App\User;
 use Tests\TestCase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -14,25 +13,10 @@ class LoginControllerTest extends TestCase
      *
      * @return void
      */
-    public function testLogin()
+    public function testExample()
     {
-        assert(User::all()->count(), 0);
-        $user = factory(User::class, 'user')->create(
-          ['password' => bcrypt($password = 'some_password'),]
-        );
-        assert(User::all()->count(), 1);
+        $response = $this->get('/');
 
-        $response = $this
-          ->actingAs($user)
-          ->json('POST', '/api/v1/auth/login',
-          [
-            'email' => $user->email,
-            'password' => $password
-          ]
-        );
-        $response
-            ->assertStatus(200)
-            ->assertSee('success')
-            ->assertSee('token');
+        $response->assertStatus(200);
     }
 }
