+++ tests/Feature/Http/Controllers/AuthControllerTest.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Feature\Http\Controllers;
 
+use App\User;
 use Tests\TestCase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -9,14 +10,29 @@ use Illuminate\Foundation\Testing\RefreshDatabase;
 class AuthControllerTest extends TestCase
 {
     /**
-     * A basic feature test example.
+     * Test API Login endpoint.
      *
      * @return void
      */
-    public function testExample()
+    public function testLogin()
     {
-        $response = $this->get('/');
+        assert(User::all()->count(), 0);
+        $user = factory(User::class, 'user')->create(
+          ['password' => bcrypt($password = 'some_password'),]
+        );
+        assert(User::all()->count(), 1);
 
-        $response->assertStatus(200);
+        $response = $this
+          ->actingAs($user)
+          ->json('POST', '/api/v1/auth/login',
+          [
+            'email' => $user->email,
+            'password' => $password
+          ]
+        );
+        $response
+            ->assertStatus(200)
+            ->assertSee('success')
+            ->assertSee('token');
     }
 }
